fix: stop calling runMigrations() on every CBT request + fix leaderboard ROUND type
- Removed runMigrations() call from CBT questions, exams, and health
  endpoints. The schema_migrations table doesn't exist in the DB and
  the migration files reference tenants(id) which also doesn't exist,
  causing every CBT request to 500. Tables are already created directly.
- Fixed homeroom leaderboard ROUND() call: PostgreSQL ROUND(x, n) requires
  numeric type, not float8. Cast to ::numeric before rounding.

+++ api/tenant/_lib/attendance.ts
@@ -1131,7 +1131,7 @@ export async function calculateHomeroomLeaderboard(
         SUM(CASE WHEN status = 'present' THEN 1 ELSE 0 END) as present_count,
         COUNT(*) as total_records,
         ROUND(
-          (SUM(CASE WHEN status = 'present' THEN 1 ELSE 0 END)::float / NULLIF(COUNT(*), 0)) * 100,
+          ((SUM(CASE WHEN status = 'present' THEN 1 ELSE 0 END)::numeric / NULLIF(COUNT(*), 0)) * 100)::numeric,
           1
         ) as rate
        FROM attendance_records
